[WT-365]:(fix) Check if indexedDB is available for use in browser

+++ src/app/database/services/database.service/indexed-db.service.ts
@@ -1,3 +1,4 @@
+import browserService, { Browser } from 'app/core/services/browser.service';
 import * as idb from 'idb';
 import { AppDatabase, DatabaseService } from '.';
 
@@ -21,13 +22,20 @@ const open = (name: string, version?: number): Promise<idb.IDBPDatabase<AppDatab
   });
 };
 
+const indexedDBIsAvailable = async () =>
+  !(await browserService.isBrowser({ browser: Browser.Firefox, incognito: true }));
 const indexedDBService: DatabaseService = (databaseName, databaseVersion) => ({
+  isAvailable: indexedDBIsAvailable,
   put: async (collectionName, key, value) => {
+    const available = await indexedDBIsAvailable();
+    if (!available) return undefined;
     const db = await open(databaseName, databaseVersion);
     await db.put(collectionName, value, key);
     db.close();
   },
   get: async (collectionName, key) => {
+    const available = await indexedDBIsAvailable();
+    if (!available) return undefined;
     const db = await open(databaseName, databaseVersion);
     const content = await db.get(collectionName, key);
     db.close();
@@ -36,17 +44,23 @@ const indexedDBService: DatabaseService = (databaseName, databaseVersion) => ({
   },
   clear: () => idb.deleteDB(databaseName),
   delete: async (collectionName, key) => {
+    const available = await indexedDBIsAvailable();
+    if (!available) return undefined;
     const db = await open(databaseName, databaseVersion);
     await db.delete(collectionName, key);
     db.close();
   },
   getAll: async (collectionName) => {
+    const available = await indexedDBIsAvailable();
+    if (!available) return undefined;
     const db = await open(databaseName, databaseVersion);
     const data = await db.getAll(collectionName);
     db.close();
     return data;
   },
   getAllFromIndex: async (collectionName, indexName, key) => {
+    const available = await indexedDBIsAvailable();
+    if (!available) return undefined;
     const db = await open(databaseName, databaseVersion);
     const data = await db.getAllFromIndex(collectionName, indexName as never, IDBKeyRange.only(key));
     db.close();
